* Optimize processing engine and encoding * Making the send interval and batch size configurable using environment variables, so that new or updated engine configuration written into the jobs database have the configurable settings, or the default if not configured. * Setting the variables to 0 or negative will configure the properties in the engine configuration to be null.

+++ components/opc-ua/src/Microsoft.Azure.IIoT.OpcUa.Edge.Publisher/src/IMessageTrigger.cs
@@ -28,13 +28,13 @@ namespace Microsoft.Azure.IIoT.OpcUa.Edge.Publisher {
         /// The number of all monitored items value changes
         /// that have been invoked by this message source.
         /// </summary>
-        int ValueChangesCount { get; }
+        long ValueChangesCount { get; }
 
         /// <summary>
-        /// The number of all dataChange Notifications 
+        /// The number of all dataChange Notifications
         /// that have been invoked by this message source.
         /// </summary>
-        int DataChangesCount { get; }
+        long DataChangesCount { get; }
 
         /// <summary>
         /// Writer events
